bug:1,辅料不能复制；2，面料名称第一次不能保存；3，待办任务增加反签收功能

+++ one/src/main/java/com/skysport/core/action/BaseAction.java
@@ -56,7 +56,7 @@ public abstract class BaseAction<T> {
      * @param draw
      * @return <K, V> Map<K, V>
      */
-    public Map<String, Object> buildSearchJsonMap(List<T> results, long recordsTotal, int recordsFiltered, int draw) {
+    public Map<String, Object> buildSearchJsonMap(List<T> results, long recordsTotal, long recordsFiltered, int draw) {
         Map<String, Object> info = new HashMap<String, Object>();
         info.put("data", results);
         info.put("recordsTotal", recordsTotal);
